pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
homepage = "https://github.com/radiantlogicinc/fastworkflow"
repository = "https://github.com/radiantlogicinc/fastworkflow"
# documentation = "https://fastworkflow.ai/readthedocs"

[tool.poetry]
name = "fastworkflow"
version = "2.17.32"
description = "A framework for rapidly building large-scale, deterministic, interactive workflows with a fault-tolerant, conversational UX"
authors = ["Dhar Rawal <drawal@radiantlogic.com>"]
license = "Apache-2.0"
readme = "README.md"
keywords = ["fastworkflow", "ai", "workflow", "llm", "openai"]
include = [
    "fastworkflow/examples/fastworkflow.env",
    "fastworkflow/examples/fastworkflow.passwords.env",
    "fastworkflow/_commands/.gitkeep"
]
exclude = [
    "**/___command_info",
    "**/___workflow_contexts",
    "./.env",
    "**/simple_workflow_template/tests"
]

[tool.poetry.scripts]
fastworkflow = "fastworkflow.cli:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pydantic = "^2.9.2"
speedict = "^0.3.12"
python-dotenv = "^1.0.1"
scikit-learn = "^1.6.1"
transformers = "^4.48.2"
mmh3 = "^5.1.0"
python-levenshtein = "^0.27.1"
colorama = "^0.4.6"
dspy = "^3.0.1"
torch = "^2.7.1"
prompt_toolkit = "^3.0.43"
libcst = "^1.8.2"
openai = ">=2.8.0"
litellm = {extras = ["proxy"], version = "^1.81.4"}
uvicorn = {version = ">=0.31.1,<0.32.0", optional = true}
fastapi = {version = ">=0.120.1", optional = true}
fastapi-mcp = {version = "^0.4.0", optional = true}
pyjwt = {extras = ["crypto"], version = ">=2.4.0", optional = true}  # Replaces python-jose (CVE-2024-23342 in ecdsa dep)
datasets = {version = "^4.0.0", optional = true}

# Security floor constraints for transitive dependencies
# These ensure users get patched versions even though these are not direct deps
orjson = ">=3.9.15"      # CVE-2024-27454 (loads recursion) - NOTE: dumps recursion DoS <=3.11.4, no upstream patch
urllib3 = ">=2.6.0"      # CVE-2025-66418, CVE-2025-66471 (decompression chain)
aiohttp = ">=3.13.3"     # CVE-2025-69223 + 7 others (zip bomb, DoS)
starlette = ">=0.49.1"   # CVE-2025-62727, CVE-2025-54121 (Range header DoS)
pyasn1 = ">=0.6.2"       # CVE-2026-23490 (decoder DoS)
filelock = ">=3.20.1"    # CVE-2025-68146 (TOCTOU symlink)
pynacl = ">=1.6.2"       # CVE-2025-69277 (libsodium validation)
# NOTE: fastapi-sso CSRF vuln (needs >=0.19.0) blocked by litellm pinning to <0.17.0 - upstream issue

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
pylint = "^3.3.1"
mypy = "^1.13.0"
bandit = "^1.7.10"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
requests-mock = "^1.12.1"
pytest-cov = "^6.2.1"

# poetry install --with aws if you are using AWS Bedrock for LLM or LLM_AGENT
[tool.poetry.group.aws.dependencies]
boto3 = {version = ">=1.36.0", optional = true}

[tool.poetry.extras]
training = ["datasets"]
fastapi = ["uvicorn", "fastapi", "fastapi-mcp", "pyjwt"]
